Getting too many test files in root, move to __tests__
